+++ lib/document_scanner_flutter.dart
@@ -1,19 +1,26 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:document_scanner_flutter/screens/pdf_generator_gallery.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-enum ScannerFileSource { CAMERA, GALLERY }
+import 'configs/configs.dart';
 
 class DocumentScannerFlutter {
   static MethodChannel get _channel =>
       const MethodChannel('document_scanner_flutter');
 
-  static Future<File?> _scanDocument(ScannerFileSource source) async {
-    String? path =
-        await _channel.invokeMethod(describeEnum(source).toLowerCase());
+  static Future<File?> _scanDocument(
+      ScannerFileSource source, Map<dynamic, String> androidConfigs) async {
+    Map<String, String?> finalAndroidArgs = {};
+    for (var entry in androidConfigs.entries) {
+      finalAndroidArgs[describeEnum(entry.key)] = entry.value;
+    }
+
+    String? path = await _channel.invokeMethod(
+        describeEnum(source).toLowerCase(), finalAndroidArgs);
     if (path == null) {
       return null;
     } else {
@@ -24,10 +31,21 @@ class DocumentScannerFlutter {
     }
   }
 
+  static Future<File?> launchForPdf(BuildContext context,
+      {Map<dynamic, String> androidConfigs = const {}}) async {
+    Future<File?>? launchWrapper() {
+      return launch(context, androidConfigs: androidConfigs);
+    }
+
+    return await Navigator.push<File>(context,
+        MaterialPageRoute(builder: (_) => PdfGeneratotGallery(launchWrapper)));
+  }
+
   static Future<File?>? launch(BuildContext context,
-      {ScannerFileSource? source}) {
+      {ScannerFileSource? source,
+      Map<dynamic, String> androidConfigs = const {}}) {
     if (source != null) {
-      return _scanDocument(source);
+      return _scanDocument(source, androidConfigs);
     }
     return showModalBottomSheet<File>(
         context: context,
@@ -39,15 +57,19 @@ class DocumentScannerFlutter {
                     leading: new Icon(Icons.camera_alt),
                     title: new Text('Camera'),
                     onTap: () async {
-                      Navigator.pop(context,
-                          await _scanDocument(ScannerFileSource.CAMERA));
+                      Navigator.pop(
+                          context,
+                          await _scanDocument(
+                              ScannerFileSource.CAMERA, androidConfigs));
                     }),
                 new ListTile(
                   leading: new Icon(Icons.image_search),
                   title: new Text('Photo Library'),
                   onTap: () async {
-                    Navigator.pop(context,
-                        await _scanDocument(ScannerFileSource.GALLERY));
+                    Navigator.pop(
+                        context,
+                        await _scanDocument(
+                            ScannerFileSource.GALLERY, androidConfigs));
                   },
                 ),
               ],